umip: umip_test2: use a long type for test mask with register operands
We use a mask to determine what parts of the returned results changed.
When operands are registers, instructions can return up to 64-bit values.
Thus, it is not sufficient to have a 16-bit mask.

+++ meta-luv/recipes-core/umip/files/umip_test2.c
@@ -223,7 +223,7 @@ static unsigned long get_mask(int op_size) {
 static int test_str(void)
 {
 	unsigned long val;
-	unsigned short mask = 0xffff;
+	unsigned long mask = 0xffff;
 
 	pr_info("====Checking STR. Expected value: [0x%x]====\n", expected_tr);
 	pr_info("==Tests for register operands==\n");
@@ -240,7 +240,7 @@ static int test_str(void)
 static int test_smsw(void)
 {
 	unsigned long val;
-	unsigned short mask = 0xffff;
+	unsigned long mask = 0xffff;
 
 	pr_info("====Checking SMSW. Expected value: [0x%x]====\n", expected_msw);
 	pr_info("==Tests for register operands==\n");
@@ -256,7 +256,7 @@ static int test_smsw(void)
 static int test_sldt(void)
 {
 	unsigned long val;
-	unsigned short mask = 0xffff;
+	unsigned long mask = 0xffff;
 
 	pr_info("====Checking SLDT. Expected value: [0x%x]====\n", expected_ldt);
 	pr_info("==Tests for register operands==\n");
